+++ src/helpers/parseWinPath.ts
@@ -1,3 +1,15 @@
+import { isWindows } from '../utils/isWindows';
+
 export const parseWinPath = (path: string | undefined): string => {
-  return path?.split(`\\`).join(`/`) || '';
+  path = path?.split(`\\`).join(`/`) || '';
+
+  if (isWindows()) {
+    // Check if path starts with a drive letter (e.g., "C:\")
+    if (/^[a-zA-Z]:\\/.test(path)) {
+      // Convert to lowercase drive letter
+      path = path.charAt(0).toLowerCase() + path.slice(1);
+    }
+  }
+
+  return path;
 };